Refactor components and hooks
