+++ genai-cookbook/app/cookbook/[recipe]/page.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 export default function Page({ params }: { params: { recipe?: string } }) {
     return <>{params.recipe && <div>Demo of {params.recipe}</div>}</>
 }
