+++ genai-cookbook/components/Navbar.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import { AppShell, Group, ScrollArea, Stack, Text } from '@mantine/core'
 import { IconChevronRight } from '@tabler/icons-react'
-import { useCookbook } from '@/hooks'
+import { useCookbook } from '@/hooks/useCookbook'
 import { cookbookRoute } from '@/lib/constants'
 import { iconStroke } from '@/lib/theme'
 import { useSelectedLayoutSegment } from 'next/navigation'
